backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["ArtemBagin <artem.bagin@vk.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.1"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
passlib = "^1.7.4"
asyncpg = "^0.29.0"
python-jose = "^3.3.0"
bcrypt = "^4.0.1"
uvicorn = "^0.24.0.post1"
greenlet = "^3.0.1"



[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
